fixed Login and Logout Features, move img and css to assets folder
